refactor(reader): extract useStudentSummaryReader hook
Move all state, effects, queries, mutations and derived values out of
StudentSummaryReader.tsx into a dedicated hook. The component becomes
thinner shell (716 -> 464 LOC) and the hook owns the orchestration
(339 LOC). No behavior change; StickyNotesPanel integration preserved.

Part 1/4 of refactor/student-summary-reader-split (issue #423).
